[147] Fixed some wrong spell and sound IDs in the Maiden of Virtue's script and tweaked her spell timers.
Fixed some wrong spell and sound IDs in the Maiden of Virtue's script and tweaked her spell timers.
Most of Prince Malchezaar's script is rewritten, now the targeting should be better. Also implemented some of the missing effects.
Added a possibility to use positions when calling SelectUnit with the attribute SELECT_TARGET_RANDOM. In that case, a random unit is selected excluding the first few ordered by threat, few being the number you pass for position. That's useful for abilities that target a random unit except the tank for example.

git-svn-id: https://scriptdev2.svn.sourceforge.net/svnroot/scriptdev2@147 5f9c896b-1e26-0410-94da-f77f675e2462