[2130] AIInfo command support, requires MaNGOS Rev [11526]
* Added for EscortAI and for ScriptedAI
* Added Thermaplugg example

Note for devs: This AIInformation can be used for more complicated scripts to add some usable debug output, which can be used on-the-fly when something went wrong within the script
This is also suggested to be used within Bug-reports (if information provided).
Also remark to request/ provide AIInformation output for scripts, which sometimes bug.