[2244] Implement helper class to handle dialogues
* Add structs DialogueEntry and DialogueEntryTwoSide to define the structure of an dialogue
DialogueEntry is What is spoken by Whom and When
DialogueEntryTwoSide is What is speaken by Whom or WhatAlternative is spoken by WhomAlternatively and When

* Add helper class DialogueHelper which does the actual work of an dialogue.
Either an object of this class must be used as member where a dialogue is to be used,
or this class can be used with multi-inheriting to be able to react when a dialogue"event" happend. (Use the multi-inheriting only when needed)

Thanks to crackm and vladimir for helping out with programming techniques!