ICC, Instance - Add class for instance script
In this commit I add the base layout for a SD2 instance script.
This should be created as a subclass of ScriptedInstance, and implemented with a header declaration.

The reason for the header declaration is, that this is required to create some special function in the instance class and use them from boss scripts.

The script will be hooked to the ScriptName 'instance_icecrown_citadel'
It will be added to SQL in the next commit.